+++ mp4decrypt/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2025-06-10
+
+### Changed
+
+- Now reference to hashmap is required.
+- The `Error` struct is returned as error variant.
+
 ## [0.4.2] - 2023-06-12
 
 ### Added
@@ -62,7 +69,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2022-08-31
 
-[Unreleased]: https://github.com/clitic/vsd/compare/mp4decrypt-v0.4.2...HEAD
+[Unreleased]: https://github.com/clitic/vsd/compare/mp4decrypt-0.5.0...HEAD
+[0.5.0]: https://github.com/clitic/vsd/compare/mp4decrypt-v0.4.2...mp4decrypt-0.5.0
 [0.4.2]: https://github.com/clitic/vsd/compare/mp4decrypt-v0.4.1...mp4decrypt-v0.4.2
 [0.4.1]: https://github.com/clitic/vsd/compare/mp4decrypt-v0.4.0...mp4decrypt-v0.4.1
 [0.4.0]: https://github.com/clitic/mp4decrypt/compare/v0.3.1...HEAD